+++ lib/src/modular_base.dart
@@ -6,6 +6,8 @@ import 'package:flutter_modular/src/routers/router.dart';
 
 import 'interfaces/child_module.dart';
 import 'interfaces/route_guard.dart';
+import 'navigator/modular_navigator.dart';
+import 'navigator/modular_navigator_interface.dart';
 
 _debugPrintModular(String text) {
   if (Modular.debugMode) {
@@ -39,62 +41,10 @@ class Modular {
 
   static init(ChildModule module) {
     _initialModule = module;
-    if (Modular.debugMode) {
-      _printRouters();
-    }
     bindModule(module, "global==");
   }
 
-  static _printRouters() {
-    List<Type> typesCheckds = [];
-    List<String> paths = [];
-    _printRoutersModule(_initialModule, '/', paths, typesCheckds);
-    paths.sort((preview, actual) => preview
-        .split(' => ')[0]
-        .length
-        .compareTo(actual.split(' => ')[0].length));
-    int sizedPath = paths[paths.length - 1].split(' => ')[0].length;
-    paths = paths.map((p) {
-      List<String> split = p.split(' => ');
-      int sizeLocalPath = split[0].length;
-      if (sizedPath >= sizeLocalPath) {
-        String spaces =
-            List.generate(sizedPath - sizeLocalPath, (index) => ' ').join('');
-        String path = split[0];
-        if (path != '/' && path[path.length - 1] == '/') {
-          path = path.substring(0, path.length - 1);
-        }
-        return "$path$spaces => ${split[1]}";
-      }
-      return p;
-    }).toList();
-
-    debugPrint('\n*** Modular Routers ***\n');
-    paths.forEach(print);
-    debugPrint("\n*****\n");
-  }
-
-  static _printRoutersModule(ChildModule module, String initialPath,
-      List<String> paths, List<Type> typesCheckds) {
-    typesCheckds.add(module.runtimeType);
-    for (var router in module.routers.where((router) => router.child != null)) {
-      String page = router.child.runtimeType
-          .toString()
-          .replaceFirst('(BuildContext, ModularArguments) => ', '');
-      String path = "$initialPath${router.routerName}".replaceFirst('//', '/');
-      paths.add('$path => $page');
-    }
-
-    bool _condition(router) => (router.module != null &&
-        !typesCheckds.contains(router.module.runtimeType));
-
-    for (var router in module.routers.where(_condition)) {
-      _printRoutersModule(
-          router.module, router.routerName, paths, typesCheckds);
-    }
-  }
-
-  static NavigatorState get to {
+  static IModularNavigator get to {
     assert(
         _navigatorKey != null, '''Add Modular.navigatorKey in your MaterialApp;
 
@@ -104,7 +54,7 @@ class Modular {
 
 .
       ''');
-    return _navigatorKey.currentState;
+    return ModularNavigator(_navigatorKey.currentState);
   }
 
   @visibleForTesting